created cart service, added feedback, added ang feature
